models/ensemble_cpr/requirements.txt
xgboost
scikit-learn
pandas
google-cloud-aiplatform
